Renamed to backup_remove_command.rs.
