+++ src/backup_remove_command.rs
